Installation des tables authors et publishers

+++ priv/repo/seeds.exs
@@ -33,52 +33,54 @@ end
 
 # ==== Pages et Pages Locales ====
 
+if false do ### METTRE À TRUE POUR TRAITER
 
+  Repo.delete_all(PageLocale)
+  Repo.delete_all(Page)
 
-Repo.delete_all(PageLocale)
-Repo.delete_all(Page)
+  path_folder = Path.join(["assets","pages","fr"])
+  File.ls!(path_folder)
+  |> Enum.filter(fn path -> String.ends_with?(path, ".phil") end)
+  |> Enum.each(fn path ->
+    slug = Path.rootname(path)
+    fullpath = Path.join([path_folder, path])
+    page = Repo.insert!(%Page{
+      slug: slug, 
+      template: "plain_page", 
+      status: 5
+    })
 
-path_folder = Path.join(["assets","pages","fr"])
-File.ls!(path_folder)
-|> Enum.filter(fn path -> String.ends_with?(path, ".phil") end)
-|> Enum.each(fn path ->
-  slug = Path.rootname(path)
-  fullpath = Path.join([path_folder, path])
-  page = Repo.insert!(%Page{
-    slug: slug, 
-    template: "plain_page", 
-    status: 5
-  })
+    # On récupère les données dans l'entête du fichier .phil
+    pdata = PhilHtml.to_data(fullpath, [evaluation: false, to_file: false, no_header: true])
+    
+    [title, summary] =
+      if Map.has_key?(pdata.metadata, :title) do
+        tit = pdata.metadata.title || "#{slug à nommer}"
+        sum = pdata.metadata.summary || "-- RÉSUMÉ À DÉFINIR ---"
+        [tit, sum]
+      else
+        IO.puts [IO.ANSI.red() <> "Il faut définir l'entête du fichier" <> IO.ANSI.reset()]
+        ["#{slug} À RENOMMER", "-- RÉSUMÉ À DÉFINIR ---"]
+      end
 
-  # On récupère les données dans l'entête du fichier .phil
-  pdata = PhilHtml.to_data(fullpath, [evaluation: false, to_file: false, no_header: true])
-  
-  [title, summary] =
-    if Map.has_key?(pdata.metadata, :title) do
-      tit = pdata.metadata.title || "#{slug à nommer}"
-      sum = pdata.metadata.summary || "-- RÉSUMÉ À DÉFINIR ---"
-      [tit, sum]
-    else
-      IO.puts [IO.ANSI.red() <> "Il faut définir l'entête du fichier" <> IO.ANSI.reset()]
-      ["#{slug} À RENOMMER", "-- RÉSUMÉ À DÉFINIR ---"]
-    end
+    html_path = Path.join([path_folder, "xhtml", "#{slug}.html"])
+    content = if File.exists?(html_path) do
+      File.read!(html_path)
+    else "À TRANSVASER" end
 
-  html_path = Path.join([path_folder, "xhtml", "#{slug}.html"])
-  content = if File.exists?(html_path) do
-    File.read!(html_path)
-  else "À TRANSVASER" end
+    locpage = Repo.insert!(%PageLocale{
+      page_id:      page.id,
+      locale:       "fr",
+      status:       5,
+      title:        title,
+      raw_content:  File.read!(Path.join([path_folder, path])),
+      content:      content,
+      summary:      summary
+    })
+    IO.puts "-> #{slug} OK"
+  end)
 
-  locpage = Repo.insert!(%PageLocale{
-    page_id:      page.id,
-    locale:       "fr",
-    status:       5,
-    title:        title,
-    raw_content:  File.read!(Path.join([path_folder, path])),
-    content:      content,
-    summary:      summary
-  })
-  IO.puts "-> #{slug} OK"
-end)
+end #/if il faut traiter les pages
 
 
 # Ajouter quelques éditeurs connus
